+++ Game/CGame/cg_local.h
@@ -1359,7 +1359,7 @@ void		trap_SetUserCmdValue( int stateValue, float sensitivityScale, int tierStat
 void		testPrintInt( char *string, int i );
 void		testPrintFloat( char *string, float f );
 
-int			trap_MemoryRemaining( void );
+size_t		trap_MemoryRemaining( void );
 void		trap_R_RegisterFont(const char *fontName, int pointSize, fontInfo_t *font);
 qboolean	trap_Key_IsDown( int keynum );
 int			trap_Key_GetCatcher( void );
